Refactor accounting modules to use Decimal instead of float for monetary values
- Updated balance_sheet.py to convert Decimal128 values to Decimal for arithmetic operations.
- Modified converted_summary_class.py to use Decimal for currency fields and added a method to create instances from CryptoConv.
- Enhanced ledger_entry_class.py to convert Decimal128 to Decimal for Pydantic validation.
- Changed limit_check_classes.py to use Decimal for msats and sats fields.
- Updated profit_and_loss.py to convert Decimal128 values to Decimal for report generation.
- Added logging for error handling in accounts.py when fetching account balances.
- Adjusted users.py to handle balance_sats as Decimal, improving type safety.
- Modified balance_result.html and balance_result_javascript.html templates to format Decimal values correctly.
- Updated financial_reports templates to ensure proper formatting of monetary values as Decimal.
- Refactored crypto_conversion.py to use Decimal for all monetary fields and conversions.
- Enhanced general_purpose_funcs.py to convert Decimal instances to appropriate MongoDB types.
- Updated custom_json_data.py to use Decimal for msats field.
- Modified pydantic_helpers.py to convert Decimal128 fields to Decimal in item dictionaries.
- Adjusted tests to compare Decimal values correctly, ensuring precision in assertions.
- Created frontend_classes.py as a new module for future frontend-related classes.

+++ src/v4vapp_backend_v2/hive_models/custom_json_data.py
@@ -1,3 +1,4 @@
+from decimal import Decimal
 from typing import Any, Dict, List, Type, Union
 
 from pydantic import BaseModel, ConfigDict, Field
@@ -66,7 +67,7 @@ class KeepsatsTransfer(BaseModel):
         ge=0,
         description="The amount of sats being transferred. Not needed if we are sending a fixed amount invoice, used if we are using a lightning address or zero value invoice (used as an upper limit sometimes)",
     )
-    msats: int | None = Field(
+    msats: Decimal | None = Field(
         None,
         ge=0,
         description=(
